Background_Image_Swap/Background_Image_Swap/requirements.txt
streamlit
rembg
pillow
numpy
onnxruntime

To run app use command: 
1. streamlit run app.py --server.enableCORS false --server.enableXsrfProtection false
2. streamlit run app2.py --server.enableCORS false --server.enableXsrfProtection false